# Improve speed of bitwiseNOT.