--- a/testsrc/test262.properties
+++ b/testsrc/test262.properties
@@ -2376,6 +2376,8 @@ language/expressions/arrow-function
     ! syntax/early-errors/asi-restriction-invalid.js
     ! syntax/early-errors/use-strict-with-non-simple-param.js
 
+language/expressions/bitwise-not
+
 language/expressions/delete
     ! 11.4.1-3-a-1-s.js
     ! 11.4.1-5-a-1-s.js