@gbrail 2021-04-03
This is an interesting optimization. Thanks!